requirements.txt
# Core FastAPI dependencies (existing)
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
psycopg2-binary==2.9.9
pydantic-settings==2.0.1
python-dotenv==1.0.1 # For .env file support
gunicorn==21.2.0 # For production deployment
alembic==1.12.0
apscheduler==3.10.1
boto3==1.34.74 # For AWS Timestream integration

# NEW: Payment Processing
stripe==7.8.0

# NEW: Google Services
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# NEW: SMS/Messaging
twilio==8.10.3

# NEW: HTTP Requests
requests==2.31.0
httpx==0.25.2

# NEW: Social Media APIs
# Snapchat and TikTok use standard HTTP requests, no additional packages needed

# NEW: Data Processing
python-dateutil==2.8.2
pytz==2023.3

# NEW: Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# NEW: Background Tasks
celery==5.3.4
redis==5.0.1

# NEW: Monitoring
sentry-sdk[fastapi]==1.38.0

# NEW: Testing
pytest-cov==4.1.0
factory-boy==3.3.0